dkim: signer: log info instead of die'ing when missing domain
for mail locally generated by PMG the signing sources
(envelope-sender, From header) can contain only a local-part
(postmaster) or even be empty (envelope-sender).

While such mail cannot be sensibly signed, it should be treated as if
the domain is not listed in DKIM-domains - by an log message on 'info'
level instead of a `die`.
the sub with the changed behavior is only used in this module, and
sign_entity as external entry-point is only called in eval context,
resulting in a log message on level 'warn'.

so effectively this change should only reduce log-levels for DKIM
failures in these cases from 'warning' to 'info'

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>