trim Message-ID when parsing E-mail
when we currently parse emails we do not remove trailing newlines from
the message-id[0]. The consequence of this is that if you use the rule
system macro __MSGID__, there is also a newline at the end of the
string. This in turn leads to problems if you use __MSGID__ in a
ModField action and want to add something after the message ID.
One use case for this is if you want to remove or change the domain from
the MSGID for privacy reasons.

[0] MIME::Head (https://metacpan.org/pod/MIME::Head) docs explictly say
    to chomp the result to get consistentcy
Signed-off-by: Hannes Duerr <h.duerr@proxmox.com>
[S.I. remove blanks before sub-invocation, reword commit-message]
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>