schema: fix verification for transport-domain-or-nexthop type
Since pmg_verify_transport_address() is called with $noerr set, it
will not die on failure. Make sure that the verification method does
not quietly accept an invalid value in this case by making sure the
code for failure is also executed in this scenario, i.e. moving it
out of the 'else' branch.

As reported in the community forum [0], this could case issues when a
CIDR is specified instead of an IP or domain name for TLS destination
policy, which is not intended.

[0]: https://forum.proxmox.com/threads/165167/

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>