rename black/whitelist to block/welcomelist
This is the internal part of #3755 - renaming all occurrences
of blacklist to blocklist and whitelist to welcomelist.

The choice of welcome/block is following SpamAssassin, and while
sounding a bit clumsy has the advantage that the acronyms WL/BL still
match.

done with `s/hite/elcome/` and `s/lack/lock/` (and manual selection of
the respective changes).

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>