pmgtunnel: adapt to read_etc_network_interfaces using run_command
the changes in pve-common:
346ff97 ("inotify/interfaces: use 'ip link' instead of /proc/net/dev")
made the /etc/network/interfaces parser call `ip link -details` via
run_command. This was the first use of a external command apart from
the ssh processes for the tunnels.

pmgtunnel changed the SIGCHLD handler before entering it's main-loop,
which in turn made run_command return an error on each invocation:
`command 'ip -details -json link show' failed: failed to execute`

The issue is not directly reproducible, as /etc/network/interfaces is
only read if the node-name does not resolv to a non-loopback ip (via
/etc/hosts or DNS).

reported in our community-forum:
https://forum.proxmox.com/threads/.169854/

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>