pmgtunnel: do not set SIGCHLD handler
Drop the SIGCHLD handling in pmgtunnel, as it's not necessary and
parts deeper in our code-base (e.g. PVE::Tools::run_command) rely on
signals not being handled by the callers.

In the case of pmgtunnel the signalhander (finish_children) simply
wait(2)'s for all children (no matter if it's a ssh-tunnel process, or
e.g. `ip link` being called to get network information), and clears
the ssh-forwarded postgres socket, schedules a (delayed) restart, and
logs an exit message for the ssh-tunnels.

all those tasks can happen a synchronously in the main loop in run as
well (especially the cleaning of the postgres socket is done directly
before running ssh anyways).

from a quick look through our perl codebase it seems that this is the
only service setting a similar SIGCHLD handler (thus the change in
reading/parsing /etc/network/interfaces did not cause issues anywhere
else).

Suggested-by: Wolfgang Bumiller <w.bumiller@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Tested-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250818180714.22846-1-s.ivanov@proxmox.com