dkim: local mail: fix `Date` header formatting under different locales
This was reported in the community forum to cause problems if the
(shell) locale was set to something different than 'C' or 'en_US.UTF-8'
[0].

Tl;dr: A different locale would cause the `Date` header to be formatted
according to that locale, instead of following the RFC2822 format.

E.g. under the locale `de_DE.UTF-8`, the header would look like this:

  Date: Mo, 25 Aug 2025 00:01:04 +0200

instead of the correct format of:

  Date: Mon, 25 Aug 2025 00:01:04 +0200

`perldoc POSIX` documents the following under strftime:

> For example, the specifiers `aAbBcpZ` change according to the locale
> settings of the user, [..]

Tested this by running the following, before and after:

  export LC_TIME=de_DE.UTF-8
  unset LC_ALL # ensure LC_TIME is used
  pmgreport

.. and verifying that the date header is now correct.

[0] https://forum.proxmox.com/threads/datumsprobleme-mit-den-pmg-reports.170245

Fixes: 77cfddd ("dkim: local mail: add Date header")
Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>