d/postinst: ensure that clamav-freshclam is enabled
recent versions of clamav do not enable clamav-freshclam by default
see https://salsa.debian.org/clamav-team/clamav/-/merge_requests/7 for
a bit of background.

As we want clamav-freshclam to run for PMG installations - enable it
in the postinst script, unless it is masked (to leave an escape-hatch
for admins).

Instead of explicitly checking for the service being masked (there is
no `is-masked` verb for `systemctl` - just try to enable it in any
case - if the unit is masked a (not scary looking) message is printed
for the user.

reported in our community forum:
see https://forum.proxmox.com/threads/.171250/

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>