pmg8to9: add context and adapt deprecated who object entry warning
The warning stems from #5972 where it was pointed out that
'fromthisdomain.com' is not reserved and suitable for examples - so
provide a bit of context to the user.

I also wanted to 'Blocklist' and 'Welcomelist' to the object groups to
check, however as we do not rename the present objectgroups and the
default ruleset does not contain '@fromthisdomain.com' addresses since
before we renamed the block/welcome list, this should not be needed.

Suggested-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
(cherry picked from commit 262d61b31d0a87c0d74217b7394902ca5ee18045)
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>