ruledb: updatedb: fix missing 'content-type:' prefix on update
The recent change to unconditionally add
"application/vnd.microsoft.portable-executable" and replace
"application/x-ms-dos-executable" with "application/x-msdownload"

missed adding the 'content-type:' prefix to the former.

This resulted in errors in the journal after upgrading:
```
unable to load rulecache : undefined object attribute: ERROR at ...
```

While testing I verified that resetting the RuleDB to defaults (`pmgdb
init` or via GUI) also fixes the issue (but this is only viable for
non-modified rulesets).

Fixes:  51f03aee ("ruledb: updatedb: automatically adapt to new .exe mime-type")
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>