utils: finalize_report: fall back to envelope to for to-header
while looking through report sending code I noticed that pmgreport (as
one of 2 utilities (the other being pmgqm) that uses finalize_report
to send mails, never provided the `pmail` data used to set the 'To'
header (a.k.a RFC5322.to).

While this seems to not have bothered anyone since 2017, when this was
introduced with:
 3dfe317a ("pmgreport: new tool to send daily system reports")
we are now reworking the report-generation quite a bit, so it makes
sense to add it now.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>