quarantine: pmgqm: utils: rename pmail_raw to pmail_plain
with the recent addition of plain-text reports we named most variables
used in the plain-text reports with the suffix _plain e.g.
`subject_plain`, as the variables without suffix (`subject`) were
html-encoded.

`pmail` is used in the reports - instead of adding another variable,
rename `pmail_raw`, which contains the same data.
`pmail_raw` was orignally introduced when we added UTF-8 support in:
 6dfa0834 ("pmgqm: handle smtputf8 data")

Technically this is a breaking change, as the variables are used by
templates, which users can override and modify.
However as  our shipped templates never used `pmail_raw` I consider
the potential of somebody using it in an override quite minimal.
With the upcoming new major the change will get mentioned in the
release-notes additionally.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>