fix #5438: api: allow admin users access to mime types endpoint
The list of mime-types recognized by the system is not really
sensitive information. The call itself reads a directory from disk,
which technically has a potential for causing load (but that should be
cached after the first read).
As it had no permission set in the API schema it was restricted to
root only.

Allowing it for all admin users is definitively OK, and as the list of
mime types is only used in what-objects, which in turn is only usable
for admins anyway, we do not gain anything by allowing access to other
roles. So, while it probably would be OK to give all authenticated
users access to this, lets error on the side of caution and only give
admins access for now until we actually get an use out of being more
permissive.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20250923092611.5058-1-s.ivanov@proxmox.com
 [TL: reword commit message a bit]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>