config: silence ENOENT when comparing SMTP filter config
If `/run/pmg-smtp-filter.cfg` doesn't yet exist when `pmgsync.service`
runs, a "No such file or directory" warning is logged. While this has
been around since a while [0], the warning might look suspicious to
some (as it did to me during testing).

Therefore, silence the warning if there's an ENOENT.

[0]: https://git.proxmox.com/?p=pmg-api.git;a=commitdiff;h=e5b31a6178118981e784b225941f1502ba644ee5

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>