8to9: clarify warning for deprecated Who-Object and provide possible action
As it was not really made clear that this is not something that was
caused by the users action, neither was it clear where this is
actually defined (e.g. Wolfgang checked /etc for references in a
config file) and it also was not documented how one can actually
remove the problematic entries.

While very seasoned PMG users might figure this out, less experienced
ones might not, and it still is nice to have more context to be able
to make a informed decision independent of the experience level.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>