fix #5966: ruledb: modify header: log action
ModField (modify header) is the only action which we currently do not
log. Originally I skipped it explicitly, expecting this to add too
many lines on an average system. However it does prevent some
situations from being debugged correctly (as laid out in #5966), and
I don't think that the argument of too many log-lines outweighs that.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251002143158.25682-1-s.ivanov@proxmox.com