api: quarantine: fix wrong backward-compatibility action
While renaming to blocklist/welcomelist we kept the old
methods in the API for backward compatibilty, and added
both options as equivalent actions.
However it seems I substituted the backward compatibility name, as
well at some point and did not notice it.

Additionally the commits of renaming and re-adding the old names for
backward compatibility were not separated cleanly (thus this
patch fixes 2 commits).

Reported in our German community forum:
https://forum.proxmox.com/threads/.173910/

Tested by reproducing with a spamreport template with the old actions,
and verified that the welcomelist action works with this patch.

Fixes: ba7a7855 ("api: quarantine: re-add black- whitelist api methods for compatibilty")
Fixes: 85981ebd ("api: quarantine: rename to block/welcomelist")
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>