utils: finalize_report: change order of plain and html parts
After 2 reports and a hint in our community forum [0] that the reports
are displayed differently since the upgrade - I checked the MIME-RFC:
https://datatracker.ietf.org/doc/html/rfc2046#section-5.1.4

which suggests to add the most preferred alternative last (so that
text/plain comes in the beginning which is nicer to view for
mail-clients which do not handle MIME at all).

I had chalked this up to my MUAs preferring text/plain where possible.

Tested this by comparing a spamreport sent with this patch applied in
* claws-mail
* mutt
* bluemind webmail
* aerc

[0] https://forum.proxmox.com/threads/.173435/

Fixes: 542f7279 ("utils: allow specifying plain and/or html for finalize_report()")
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>