api: apt: restrict package name and separate arguments from options
In order to sanitize the package names - restrict them to a smaller
set, in addition to the set allowed by the policy-manual:
https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-package
I added ':' used as a architecture separator (e.g. zlib1g:amd64).

I considered allowing also capital letters (A-Z) in that list as
I've seen a few third party packages having them in the downloadable
package-name - however none of those are present on a PMG system, and
those I did check - all had their control-file with only lower case
ascii letters in the Package field. So I left it out for now, we can
always allow for those later if they are needed.

Fixes: 785364eb ("apt: use `apt changelog` for changelog fetching")
Reported-by: Kevin Joensen <kevin@baldur.dk>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>